requirements.txt

Flask==3.0.0
Flask-Login==0.6.3
python-dotenv==1.0.1
mysql-connector-python==9.0.0
Werkzeug==3.0.4
